requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.116.1
fastjsonschema==2.21.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
future==1.0.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
kiwisolver==1.4.7
kubernetes==33.1.0
langchain==0.3.26
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langsmith==0.4.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
nbformat==5.10.4
numpy==2.0.2
oauthlib==3.3.1
onnxruntime==1.19.2
openai==1.95.1
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
posthog==5.4.0
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.2
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
zipfile36==0.1.3
zipp==3.23.0
zstandard==0.23.0
